extract API into interface Subscriber
